pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "buckets-man"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.1",
    "minio>=7.2.20",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
